+++ rust/basic_bitcoin/src/basic_bitcoin/src/util.rs
@@ -1,9 +1,3 @@
-use bitcoin::{
-    blockdata::script::Builder, hashes::Hash, Address, AddressType, OutPoint, Script, SigHashType,
-    Transaction, TxIn, TxOut, Txid,
-};
-use ic_btc_types::Utxo;
-use ic_cdk::print;
 use sha2::Digest;
 
 pub fn p2pkh_address_from_public_key(public_key: Vec<u8>) -> String {
@@ -30,145 +24,6 @@ pub fn sha256(data: Vec<u8>) -> Vec<u8> {
     let mut hasher = sha2::Sha256::new();
     hasher.update(data);
     hasher.finalize().to_vec()
-}
-
-// The signature hash type that is always used.
-const SIG_HASH_TYPE: SigHashType = SigHashType::All;
-
-// Builds a transaction that sends the given `amount` of satoshis to the `destination` address.
-pub fn build_transaction(
-    utxos: Vec<Utxo>,
-    source: Address,
-    destination: Address,
-    amount: u64,
-    fees: u64,
-) -> Result<Transaction, String> {
-    // Assume that any amount below this threshold is dust.
-    const DUST_THRESHOLD: u64 = 10_000;
-
-    // Select which UTXOs to spend. For now, we naively spend the first available UTXOs,
-    // even if they were previously spent in a transaction.
-    let mut utxos_to_spend = vec![];
-    let mut total_spent = 0;
-    for utxo in utxos.into_iter().rev() {
-        total_spent += utxo.value;
-        utxos_to_spend.push(utxo);
-        if total_spent >= amount + fees {
-            // We have enough inputs to cover the amount we want to spend.
-            break;
-        }
-    }
-
-    print(&format!("UTXOs to spend: {:?}", utxos_to_spend));
-    print(&format!(
-        "UTXO transaction id: {}",
-        Txid::from_hash(Hash::from_slice(&utxos_to_spend[0].outpoint.txid).unwrap()),
-    ));
-
-    if total_spent < amount {
-        return Err("Insufficient balance".to_string());
-    }
-
-    let inputs: Vec<TxIn> = utxos_to_spend
-        .into_iter()
-        .map(|utxo| TxIn {
-            previous_output: OutPoint {
-                txid: Txid::from_hash(Hash::from_slice(&utxo.outpoint.txid).unwrap()),
-                vout: utxo.outpoint.vout,
-            },
-            sequence: 0xffffffff,
-            witness: Vec::new(),
-            script_sig: Script::new(),
-        })
-        .collect();
-
-    let mut outputs = vec![TxOut {
-        script_pubkey: destination.script_pubkey(),
-        value: amount,
-    }];
-
-    let remaining_amount = total_spent - amount - fees;
-
-    if remaining_amount >= DUST_THRESHOLD {
-        outputs.push(TxOut {
-            script_pubkey: source.script_pubkey(),
-            value: remaining_amount,
-        });
-    }
-
-    Ok(Transaction {
-        input: inputs,
-        output: outputs,
-        lock_time: 0,
-        version: 2,
-    })
-}
-
-/// Sign a bitcoin transaction given the private key and the source address of the funds.
-///
-/// Constraints:
-/// * All the inputs are referencing outpoints that are owned by `src_address`.
-/// * `src_address` is a P2PKH address.
-pub async fn sign_transaction(
-    mut transaction: Transaction,
-    src_address: Address,
-    public_key: Vec<u8>,
-) -> Transaction {
-    // Verify that the address is P2PKH. The signature algorithm below is specific to P2PKH.
-    match src_address.address_type() {
-        Some(AddressType::P2pkh) => {}
-        _ => panic!("This demo supports signing p2pkh addresses only."),
-    };
-
-    let txclone = transaction.clone();
-
-    for (index, input) in transaction.input.iter_mut().enumerate() {
-        let sighash =
-            txclone.signature_hash(index, &src_address.script_pubkey(), SIG_HASH_TYPE.as_u32());
-
-        let signature = crate::ecdsa::sign_with_ecdsa(sighash.to_vec()).await;
-
-        let r: Vec<u8> = if signature[0] & 0x80 != 0 {
-            // r is negative. Prepend a zero byte.
-            let mut tmp = vec![0x00];
-            tmp.extend(signature[..32].to_vec());
-            tmp
-        } else {
-            // r is positive.
-            signature[..32].to_vec()
-        };
-
-        let s: Vec<u8> = if signature[32] & 0x80 != 0 {
-            // s is negative. Prepend a zero byte.
-            let mut tmp = vec![0x00];
-            tmp.extend(signature[32..].to_vec());
-            tmp
-        } else {
-            // s is positive.
-            signature[32..].to_vec()
-        };
-
-        // Convert signature to DER.
-        let der_signature: Vec<u8> = vec![
-            vec![0x30, 4 + r.len() as u8 + s.len() as u8, 0x02, r.len() as u8],
-            r,
-            vec![0x02, s.len() as u8],
-            s,
-        ]
-        .into_iter()
-        .flatten()
-        .collect();
-
-        let mut sig_with_hashtype = der_signature;
-        sig_with_hashtype.push(SIG_HASH_TYPE.as_u32() as u8);
-        input.script_sig = Builder::new()
-            .push_slice(sig_with_hashtype.as_slice())
-            .push_slice(public_key.as_slice())
-            .into_script();
-        input.witness.clear();
-    }
-
-    transaction
 }
 
 #[test]
